Implement APIs for controlling status bar theming (edge-to-edge, color) to adapt with Android 15 behavior changes
